+++ Tests/ConsoleTests/CriticalSelectionTests.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace ConsoleTests
+{
+    static class CriticalSelectionTests
+    {
+        public static void Start()
+        {
+
+        }
+    }
+}
